Revert "feat(python): Add `replace_all` expression to complement `replace` (#…"
This reverts commit 3ac1c40dbcc05041d18ad6ae162a7e246a944e21.
