+++ py-polars/docs/source/reference/expressions/modify_select.rst
@@ -44,7 +44,6 @@ Manipulation/selection
     Expr.reinterpret
     Expr.repeat_by
     Expr.replace
-    Expr.replace_all
     Expr.reshape
     Expr.reverse
     Expr.rle
